app/requirements.txt
Flask==3.1.0
Werkzeug==3.1.3
gunicorn==23.0.0
numpy==1.26.4
Pillow==10.4.0
opencv-python-headless==4.10.0.84
tensorflow==2.18.1
matplotlib==3.9.2
Pillow
cryptography
flask-wtf
flask-limiter